requirements.txt
alabaster==0.7.12
alembic==1.12.1
amqp==5.3.1
apns-client==0.2.1
apns2==0.7.1
arrow==0.15.7
asgiref==3.5.2
asn1crypto==0.24.0
attrs==16.3.0
Babel==2.7.0
bcrypt==3.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
celery-progress==0.0.10
certifi==2019.6.16
cffi
chardet==3.0.4
click==8.1.3
click-completion==0.5.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography
cx-Oracle==7.2.0
decorator==4.4.2
Django
django-appconf==1.0.3
django-celery-beat==2.1.0
django-compressor==2.3
django-countries==5.5
django-debug-toolbar==2.1
django-debug-toolbar-request-history==0.1.0
django-filter==2.4.0
django-guardian==2.0.0
django-mssql-backend==2.8.1
django-postgrespool2==1.0.1
django-pyodbc==1.1.3
django-redis==4.10.0
django-rest-swagger==2.2.0
django-split-settings==1.0.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.15.2
drf-api-tracking==1.7.8
drf-nested-routers==0.91
dwebsocket==0.5.12
eventlet==0.26.1
Flask==2.2.5
Flask-Migrate==4.1.0
Flask-Minify==0.50
Flask-Login
Flask-SQLAlchemy==3.0.5
flower==0.9.3
future==0.18.2
geocoder==1.38.1
geographiclib==1.49
geopy==1.21.0
geos==0.2.2
gevent
gitdb==4.0.2
gitdb2==4.0.2
GitPython==3.0.5
googlemaps==4.1.0
greenlet==3.1.1
gunicorn==19.9.0
h2==2.6.2
hpack==3.0.0
htmlminf==0.1.13
httplib2==0.18.1
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
imagesize==1.1.0
importlib-metadata==4.11.4
importlib-resources==5.12.0
itsdangerous==2.1.2
itypes==1.1.0
Jinja2==3.1.6
jsmin==3.0.1
kombu==5.2.4
ladon==1.0.4
lesscpy==0.15.1
lxml
Mako==1.2.4
Markdown==3.1.1
MarkupSafe==2.1.5
monotonic==1.5
more-itertools==7.2.0
mysqlclient==1.4.6
Naked==0.1.31
newrelic==5.6.0.135
novo-jsobj==0.0.2
numpy
oauth2client==4.1.3
openapi-codec==1.3.2
optionaldict==0.1.1
packaging==20.3
paramiko==2.7.2
passlib==1.7.1
pdfkit==0.6.1
piexif==1.1.3
Pillow
pipenv==2018.11.26
ply==3.11
progress==1.5
prompt-toolkit==3.0.29
psutil==5.8.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.7
pyfcm==1.4.7
Pygments==2.4.2
PyJWT
pyminizip==0.2.6
PyMySQL==0.9.3
PyNaCl==1.4.0
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.2
PyPDF2==1.26.0
python-crontab==2.3.8
python-dateutil==2.8.0
# python-ldap
ldap3
python-memcached==1.59
python-redis-lock==3.7.0
pytz==2025.2
PyYAML==5.1.1
python-dotenv
qrcode==6.1
ratelim==0.1.6
raven==6.10.0
rcssmin==1.0.6
redis==3.3.6
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
rsa==4.6
Shapely==1.7.0
shellescape==3.4.1
shellingham==1.3.1
simplejson==3.16.0
six==1.12.0
smmap==3.0.1
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinx-bootstrap-theme
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyne==2.12.16
SQLAlchemy==2.0.44
sqlparse==0.3.0
timezonefinder==4.4.1
tornado==5.1.1
twilio==6.58.0
typing_extensions==4.7.1
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
vine==5.1.0
virtualenv==16.7.4
virtualenv-clone==0.5.3
wcwidth==0.2.5
Werkzeug==2.2.3
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
xmltodict==0.12.0
xxhash==3.6.0
yapf==0.29.0
zipp==3.15.0
WhiteNoise